ReSharper cleaning
