+++ Assets/Editor/GameEventEditor.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Reflection;
 using UnityEditor;
 using UnityEngine;
+using Assets.Scripts.Brett;
 
 
 namespace Lobodestroyo.Editor
