+++ Assets/Prefabs/Player.prefab
@@ -39,7 +39,7 @@ GameObject:
   - component: {fileID: 4757175153086592}
   - component: {fileID: 114805012238605440}
   m_Layer: 0
-  m_Name: Player (1)
+  m_Name: Player
   m_TagString: Player
   m_Icon: {fileID: 3936346786652291628, guid: 0000000000000000d000000000000000, type: 0}
   m_NavMeshLayer: 0
