--- Assets/Resources/Game Events/New Test Object.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 58fdf504c4095574c82b3642ad83511f
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
