--- Assets/Resources/Game Events/OnNotCaughtEvent.asset
@@ -1,13 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 43b66fe6845d54140b9f12ec77b2940a, type: 3}
-  m_Name: OnNotCaughtEvent
-  m_EditorClassIdentifier:
