+++ Assets/Resources/Game Events/Test Object.asset
@@ -9,5 +9,5 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 43b66fe6845d54140b9f12ec77b2940a, type: 3}
-  m_Name: OnCaughtEvent
+  m_Name: Test Object
   m_EditorClassIdentifier:
