+++ Assets/Scripts/Brett.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: bcc65eac55ed24f4d910194c81cbfbc3
-NativeFormatImporter:
+guid: 76ffa6ebc5e41a341b15188e405f7ee1
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
