+++ Assets/Scripts/Brett/GameEvent.cs
@@ -1,0 +1,48 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEditor;
+
+namespace Assets.Scripts.Brett
+{
+    [CreateAssetMenu]
+    public class GameEvent : ScriptableObject
+    {
+        private List<GameEventListener> listeners = new List<GameEventListener>();
+
+        public void Subscribe(GameEventListener listener)
+        {
+            listeners.Add(listener);
+        }
+
+        public void UnSubscribe(GameEventListener listener)
+        {
+            listeners.Remove(listener);
+        }
+
+        public void Raise()
+        {
+            for (int i = listeners.Count - 1; i >= 0; i--)
+            {
+                listeners[i].OnEventRaised();
+            }
+        }
+    }
+
+#if UNITY_EDITOR
+    [CustomEditor(typeof(GameEvent))]
+    public class GameEventEditor : Editor
+    {
+        public override void OnInspectorGUI()
+        {
+            base.OnInspectorGUI();
+            if (GUILayout.Button("Raise Event"))
+            {
+                var mt = target as GameEvent;
+                mt.Raise();
+            }
+        }
+    }
+#endif
+
+}
